backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector Database
pinecone-client==3.0.2

# LLM and Embeddings
google-generativeai==0.3.2  # For Gemini

# HuggingFace Embeddings
sentence-transformers==2.3.1

# PDF Processing
PyPDF2==3.0.1

# File Upload Support
python-multipart==0.0.6

# Text Chunking (standalone, no langchain dependency)
# We'll use a simple custom chunker instead

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
ruff==0.1.14
